requirements.txt
PyGithub
boto3
openai
google-generativeai
pyyaml
requests
python-dateutil
